Google Calendar Integration (#236)
Case Details page now makes on ajax call on load to the Google/Calendar/Get action method. This method hits the Google Calendar API to try to find an existing event at the appointment time. If an event does not exist, it displays a button to allow the user to create the event. If successful, the page refreshes the section and displays the event information from the existing calendar event.

TimeZones:
All events are saved with the correct timezone. Google Calendar automatically adjusts the time based on the users current location (timezone). To view the local time of the event, in the calendar, add a timezone to the calendar under google settings and set a label (eg. PST, EST, etc) and it will show them side by side.
Google uses IANA timezones and .NET uses unicode.org timezones. I also created a column for ISO timezone names.

Database changes were made to the Timezone relationship. A proper primary key was adding and foreign keys were updated accordingly.